Mental health project
The objective of this project is create a tool which function is being a support to experts on mental health. Then, this model would be able to detect statements or commentaries that are related with suicide behavior, and then inform to the experts that would initiate a process to guide this person.